requirements.txt
openai>=1.30.0
numpy>=1.26.0
tqdm>=4.66.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
ddtrace[llm]>=2.13.0,<4.0.0
